MaxBitFlipWindow Problem
You have an array of 0's and 1's. Determine a window [L,R], such that if
you flip the bits in that window, you will have maximum number of 1's in
your array, and then output this number of 1's.

For eg:
Input array: 1 0 0 1 0 0 1

Output: 6

Explanation:
If you choose a window [1,5], your array becomes,
1 1 1 0 1 1 1
which gives the total number of 1's now 6. So, your program should
output the number 6, i.e. the maximum number of 1's after choosing a
window.